requirements.txt
# python3.11.10
python-dotenv
kafka-python
requests
